Update Event type
